+++ resources/views/user/dashboard.blade.php
@@ -5,7 +5,7 @@
         <div>
 
             <div class="tac mb-2">
-                <img class="wh200 round" src="{{ auth()->user()->avatarUrl() }}" alt="Profile Photo">
+                <img class="wh-200px round" src="{{ auth()->user()->avatarUrl() }}" alt="Profile Photo">
             </div>
 
             <x-authit::user-navigation />